// frontend/src/app/generated-sources/openapi/model/ticketStatus.ts
/**
 * Ticketline API
 * Documentation for the Ticketline API
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TicketStatus { 
    reserved: Array<number>;
    purchased: Array<number>;
}